Add GPU pretraining path: mixed precision, grad accumulation, Colab notebook
train.py gains --amp {fp16,bf16} (autocast + GradScaler) and --grad_accum for
large effective batches on one GPU, plus --sample_prompt to watch coherence
emerge; corpus is stored int16 to cut RAM 4x for large datasets (CharDataset
converts slices to long). Add data/get_pretrain_data.py to fetch a corpus
(TinyStories / TinyShakespeare) and notebooks/pretrain_colab.ipynb for a
one-click ~57M-param pretrain on a free Colab T4. All CPU-safe; smoke-tested.

+++ src/dataset.py
@@ -25,7 +25,9 @@ class CharDataset(Dataset):
         return max(0, len(self.data) - self.block_size)
 
     def __getitem__(self, idx):
-        chunk = self.data[idx: idx + self.block_size + 1]
+        # Convert to int64 here: the data may be stored as int16 to save memory,
+        # but embedding lookups and cross-entropy need long indices.
+        chunk = self.data[idx: idx + self.block_size + 1].long()
         x = chunk[:-1]
         y = chunk[1:]
         return x, y
